+++ src/main/java/com/github/jgrade2/jgrade2/gradescope/GradescopeJsonException.java
@@ -1,0 +1,16 @@
+package com.github.jgrade2.jgrade2.gradescope;
+
+/**
+ * An exception for when the JSON to output to Gradescope does not match the
+ * schema it expects.
+ */
+public class GradescopeJsonException extends RuntimeException {
+
+    /**
+     * Create an instance of this exception.
+     * @param message The message for what caused the exception.
+     */
+    public GradescopeJsonException(String message) {
+        super(message);
+    }
+}
